Remove tmuxinator projects

+++ neovim/install.sh
@@ -33,10 +33,12 @@ install_plugins() {
 }
 
 install_plugins
-ln -s "${DOTFILES}/neovim/coc-settings.json" ~/.config/nvim/coc-settings.json
 
+if [[ ! -L ~/.config/nvim/coc-settings.json ]]; then
+  ln -s "${DOTFILES}/neovim/coc-settings.json" ~/.config/nvim/coc-settings.json
+fi
 
-if [[ ! -f ~/.config/nvim/lua/fzf.lua ]]; then
+if [[ ! -L ~/.config/nvim/lua/fzf.lua ]]; then
   mkdir -p ~/.config/nvim/lua
   ln -s "${DOTFILES}/neovim/fzf.lua" ~/.config/nvim/lua/fzf.lua
 fi
